// src/app/models/fontesEntidade.ts
export class FontesEntidade {
  monitoramento_gee_id: BigInteger;
  entidade_id: BigInteger;
  nm_classificacao: string;
  nm_combustivel: string;
  nm_fonte_emissao: string;
  nm_reduzido: string;
  mn_propriedade: number;
  qt_consumo: number;
  cd_unidade_calculo: string;
  cd_unidade_combustivel: string;
  qt_total_co2_bio: number;
  qt_total_co2_fossel: number;
}